+++ tests/functional_tests/test_general_formatting.py
@@ -56,6 +56,7 @@ from tests.util import check_formatting, read_test_data
         "unformatted/133_for_else.sql",
         "unformatted/134_databricks_type_hints.sql",
         "unformatted/135_star_columns.sql",
+        "unformatted/136_databricks_variant.sql",
         "unformatted/200_base_model.sql",
         "unformatted/201_basic_snapshot.sql",
         "unformatted/202_unpivot_macro.sql",